requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
docutils==0.14
falcon==1.4.1
idna==2.7
jsonschema==2.6.0
mysql-connector==2.1.6
passlib==1.7.1
pycparser==2.19
PyJWT==1.4.0
python-mimeparse==1.6.0
six==1.11.0
